+++ src/Farmer/Builders/Builders.ContainerGroups.fs
@@ -53,7 +53,7 @@ type ContainerGroupConfig =
       /// Credentials for image registries used by containers in this group.
       ImageRegistryCredentials : ImageRegistryCredential list
       /// IP address for the container group.
-      IpAddress : ContainerGroupIpAddress
+      IpAddress : ContainerGroupIpAddress option
       /// Name of the network profile for this container's group.
       NetworkProfile : ResourceName option
       /// The instances in this container group.
@@ -91,44 +91,49 @@ type ContainerGroupConfig =
         ]
 
 type ContainerGroupBuilder() =
+    member private __.AddPort (state, portType, port): ContainerGroupConfig =
+        {
+            state with IpAddress =
+                        match state.IpAddress with
+                        | Some ipAddresses ->
+                            { ipAddresses with Ports = ipAddresses.Ports.Add {| Protocol = portType; Port = port |} } |> Some
+                        | None -> { Type = IpAddressType.PublicAddress; Ports = [ {| Protocol = portType; Port = port |} ] |> Set.ofList } |> Some
+        }
+
     member __.Yield _ =
         { Name = ResourceName.Empty
           OperatingSystem = Linux
           RestartPolicy = AlwaysRestart
           Identity = ManagedIdentity.Empty
           ImageRegistryCredentials = []
-          IpAddress = { Type = PublicAddress; Ports = Set.empty }
+          IpAddress = None
           NetworkProfile = None
           Instances = []
           Volumes = Map.empty
           Tags = Map.empty }
-    member _.Run (state:ContainerGroupConfig) =
+    member this.Run (state:ContainerGroupConfig) =
         // Automatically apply all public-facing ports to the container group itself.
         state.Instances
         |> Seq.collect(fun i -> i.Ports |> Map.toSeq |> Seq.choose(function (port, PublicPort) -> Some port | _, InternalPort -> None))
-        |> Seq.fold (fun (state:ContainerGroupConfig) port ->
-            { state with
-                IpAddress =
-                    { state.IpAddress with
-                        Ports = state.IpAddress.Ports.Add {| Protocol = TCP; Port = port |} } }) state
+        |> Seq.fold (fun (state:ContainerGroupConfig) port -> this.AddPort (state, TCP, port)) state
 
-    member __.AddTcpPort(state:ContainerGroupConfig, port) = { state with IpAddress = { state.IpAddress with Ports = state.IpAddress.Ports.Add {| Protocol = TCP; Port = port |} } }
+    member this.AddTcpPort(state:ContainerGroupConfig, port) = this.AddPort (state, TCP, port)
 
     [<CustomOperation "name">]
     /// Sets the name of the container group.
-    member __.Name(state:ContainerGroupConfig, name) = { state with Name = name }
+    member _.Name(state:ContainerGroupConfig, name) = { state with Name = name }
     member this.Name(state:ContainerGroupConfig, name) = this.Name(state, ResourceName name)
     /// Sets the OS type (default Linux)
     [<CustomOperation "operating_system">]
-    member __.OsType(state:ContainerGroupConfig, os) = { state with OperatingSystem = os }
+    member _.OsType(state:ContainerGroupConfig, os) = { state with OperatingSystem = os }
     /// Sets the restart policy (default Always)
     [<CustomOperation "restart_policy">]
-    member __.RestartPolicy(state:ContainerGroupConfig, restartPolicy) = { state with RestartPolicy = restartPolicy }
+    member _.RestartPolicy(state:ContainerGroupConfig, restartPolicy) = { state with RestartPolicy = restartPolicy }
     member private _.SetIpAddress(state:ContainerGroupConfig, ipAddressType, ports) =
         { state with
             IpAddress =
                 { Type = ipAddressType
-                  Ports = ports |> Seq.map(fun (prot, port) -> {| Protocol = prot; Port = port |}) |> Set } }
+                  Ports = ports |> Seq.map(fun (prot, port) -> {| Protocol = prot; Port = port |}) |> Set } |> Some }
 
     /// Sets the IP addresss to a public address with a DNS label
     [<CustomOperation "public_dns">]
@@ -141,7 +146,7 @@ type ContainerGroupBuilder() =
     member __.NetworkProfile(state:ContainerGroupConfig, networkProfileName:string) = { state with NetworkProfile = Some (ResourceName networkProfileName) }
     /// Adds a UDP port to be externally accessible
     [<CustomOperation "add_udp_port">]
-    member __.AddUdpPort(state:ContainerGroupConfig, port) = { state with IpAddress = { state.IpAddress with Ports = state.IpAddress.Ports.Add {| Protocol = UDP; Port = port |} } }
+    member __.AddUdpPort(state:ContainerGroupConfig, port) = __.AddPort (state, UDP, port)
     /// Adds container image registry credentials for images in this container group.
     [<CustomOperation "add_registry_credentials">]
     member _.AddRegistryCredentials(state:ContainerGroupConfig, credentials) =